+++ src/components/index/OrderSelect.tsx
@@ -5,14 +5,44 @@ import {
   Select,
   useBreakpointValue,
 } from "@chakra-ui/react"
+import { useRouter } from "next/router"
 import { SortAscending } from "phosphor-react"
-import { ordering } from "./hooks/useOrder"
+import { useEffect } from "react"
+import { Guild, Hall } from "temporaryData/types"
+
+const ordering = {
+  name: (a: Guild | Hall, b: Guild | Hall) => {
+    const nameA = a.name.toUpperCase()
+    const nameB = b.name.toUpperCase()
+    if (nameA < nameB) return -1
+    if (nameA > nameB) return 1
+    return 0
+  },
+  oldest: (a: Guild | Hall, b: Guild | Hall) => a.id - b.id,
+  newest: (a: Guild | Hall, b: Guild | Hall) => b.id - a.id,
+  "most members": (a: Guild | Hall, b: Guild | Hall) =>
+    b.members?.length - a.members?.length,
+}
 
 const OrderSelect = ({ order, setOrder }) => {
   const icon = useBreakpointValue({
     base: <Icon as={SortAscending} />,
     md: false,
   })
+
+  const router = useRouter()
+
+  // Replacing the URL if ordering changes
+  useEffect(() => {
+    if (order === router.query.order) return
+
+    const newQuery = {
+      ...router.query,
+      order,
+    }
+
+    router.replace({ pathname: router.pathname, query: newQuery })
+  }, [order])
 
   return (
     <InputGroup
